+++ src/App.jsx
@@ -15,7 +15,7 @@ import Contact from './pages/contact.jsx';
 import About from './pages/about.jsx';
 import Movies from './pages/movies.jsx';
 import Booked from './pages/booking-confirmatin.jsx';
-import DetailedInfo from './detailedInfo.jsx';
+import DetailedInfo from './pages/DetailedInfo.jsx';
 import TicketPage from './pages/Ticket.jsx';
 // A React component is a function
 // it will run every time a state variable changes
