+++ src/main.jsx
@@ -3,7 +3,7 @@ import './utilities/auto-key-lists';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import DetailedInfo from './detailedInfo';
+
 
 ReactDOM.createRoot(document.querySelector('#root')).render(
   <React.StrictMode>
